+++ pages/markdown/misc/changelog.md
@@ -5,6 +5,7 @@
 
 
 * **Added Selection Mode:** Selection mode is a feature that allows you to "select" multiple flashcards when browsing a deck.  You can then perform bulk operations on the selected flashcards, like adding or removing a new tag to all of those flashcards.  You can learn more about selection mode [here](/help/selection-mode/).
+* **Added Bulk Suspending/Unsuspending/Deleting:** You can now suspend, unsuspend, or delete flashcards in bulk when searching.  This allows you to suspend all flashcards of a certain tag, and perform other related actions.
 * **Changed Min New Card Limit:** You can now choose to not learn any new flashcards per day in a deck, instead of being forced to learn at least one.
 * **Changed Update Deck Wording:** "Pull Changes" has been changed up "Update" for clarity.
 * **Added More Autonote Options:** Added more options to autonote, about how to divide the input text.